requirements.txt
anthropic==0.21.0
pillow>=10.0.0
docker>=7.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.11.1
streamlit>=1.25.0
selenium>=4.10.0
webdriver-manager>=4.0.0 